QoL for Login
Added temporary function to handle when a user is current in the password input and hits Enter, will try to signup with creds and if it already exists it'll attempt to login.

Also added ref for when user is in the email input and hits return, will focus into password input field.